#1323 Update to use fetch apis
