+++ lib/taiko.js
@@ -20,6 +20,7 @@ const { createJsDialogEventName } = require('./util');
 const inputHandler = require('./handlers/inputHandler');
 const domHandler = require('./handlers/domHandler');
 const networkHandler = require('./handlers/networkHandler');
+const fetchHandler = require('./handlers/fetchHandler');
 const pageHandler = require('./handlers/pageHandler');
 const targetHandler = require('./handlers/targetHandler');
 const runtimeHandler = require('./handlers/runtimeHandler');
@@ -329,7 +330,7 @@ const _closeBrowser = async () => {
   });
   clearContexts();
   let timeout;
-  networkHandler.resetInterceptors();
+  fetchHandler.resetInterceptors();
   if (_client) {
     // remove listeners other than JS dialog for beforeUnload on client first to stop executing them when closing
     await _client.removeAllListeners();
@@ -454,7 +455,7 @@ module.exports.switchTo = async (arg) => {
  * @returns {Promise}
  */
 module.exports.intercept = async (requestUrl, option, count) => {
-  await networkHandler.addInterceptor({
+  await fetchHandler.addInterceptor({
     requestUrl: requestUrl,
     action: option,
     count,
@@ -889,7 +890,7 @@ module.exports.goto = async (
     url = 'http://' + url;
   }
   if (options.headers) {
-    networkHandler.setHTTPHeaders(options.headers, url);
+    fetchHandler.setHTTPHeaders(options.headers, url);
   }
   options = setNavigationOptions(options);
   let response;
@@ -3516,14 +3517,14 @@ module.exports.metadata = {
  */
 module.exports.clearIntercept = (requestUrl) => {
   if (requestUrl) {
-    var success = networkHandler.resetInterceptor(requestUrl);
+    var success = fetchHandler.resetInterceptor(requestUrl);
     if (success) {
       descEvent.emit('success', 'Intercepts reset for url ' + requestUrl);
     } else {
       descEvent.emit('success', 'Intercepts not found for url ' + requestUrl);
     }
   } else {
-    networkHandler.resetInterceptors();
+    fetchHandler.resetInterceptors();
     descEvent.emit('success', 'Intercepts reset for all url');
   }
 };
