+++ test/unit-tests/clearIntercept.test.js
@@ -27,20 +27,20 @@ describe('clearIntercept', () => {
 
   it('should display success message if there are intercepts for the url', async () => {
     let validatePromise = validateEmitterEvent('success', 'Intercepts reset for url google.com');
-    let networkHandler = {
+    let fetchHandler = {
       resetInterceptor: () => true,
     };
-    taiko.__set__('networkHandler', networkHandler);
+    taiko.__set__('fetchHandler', fetchHandler);
     await taiko.clearIntercept('google.com');
     await validatePromise;
   });
 
   it('should display message if all intercepts are reset', async () => {
     let validatePromise = validateEmitterEvent('success', 'Intercepts reset for all url');
-    let networkHandler = {
+    let fetchHandler = {
       resetInterceptors: () => {},
     };
-    taiko.__set__('networkHandler', networkHandler);
+    taiko.__set__('fetchHandler', fetchHandler);
     await taiko.clearIntercept();
     await validatePromise;
   });
@@ -49,10 +49,10 @@ describe('clearIntercept', () => {
       'success',
       'Intercepts not found for url google.com',
     );
-    let networkHandler = {
+    let fetchHandler = {
       resetInterceptor: () => false,
     };
-    taiko.__set__('networkHandler', networkHandler);
+    taiko.__set__('fetchHandler', fetchHandler);
     await taiko.clearIntercept('google.com');
     await validatePromise;
   });
