+++ test/unit-tests/test-util.js
@@ -32,6 +32,7 @@ module.exports.removeFile = (filePath) => {
 };
 
 module.exports.openBrowserArgs = {
+  dumpio: true,
   args: [
     '--disable-gpu',
     '--disable-dev-shm-usage',
